Move blockListener to blocks.js; combine flyout listener
